Namespace refactoring to organize exceptions with subsequent updating of PHPUnit tests.
